requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-telegram-bot[job-queue]==20.8
langchain==0.1.9
langchain-google-genai==0.0.9
langchain-openai==0.0.8
pinecone-client==3.1.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
# httpx removed to allow auto-resolution
beautifulsoup4==4.12.3
jinja2==3.1.3
python-multipart==0.0.9
email-validator==2.1.0.post1
